add UT for investigate and modify file path-related security coding issues

+++ util/common/src/util.cpp
@@ -74,7 +74,7 @@ int64_t GetSysClockTime()
     }
 
     uint64_t totalMicroSeconds = static_cast<uint64_t>(ts.tv_sec) * TIME_CONVERSION_UNIT *
-    TIME_CONVERSION_UNIT + ts.tv_nsec / TIME_CONVERSION_UNIT;
+    TIME_CONVERSION_UNIT + static_cast<uint64_t>(ts.tv_nsec) / TIME_CONVERSION_UNIT;
 
     if (totalMicroSeconds > static_cast<uint64_t>(std::numeric_limits<int64_t>::max())) {
         MMI_HILOGE("Total time value integer overflow detected!");
